// src/components/StyledComponents/Container/interfaces.ts
export interface ContainerProps {
	children: React.ReactNode;
	width?: string;
	height?: string;
	padding?: string;
	margin?: string;
	backgroundColor?: string;
	borderRadius?: string;
	boxShadow?: string;
}

export interface StyledContainerProps {
	width?: string;
	height?: string;
	padding?: string;
	margin?: string;
	backgroundColor?: string;
	borderRadius?: string;
	boxShadow?: string;
}